Force all compiler warnings to report as errors. (#1076)
* Force all compiler warnings to report as errors.

* Fix missed errors.

* Additional Linux and Windows compiler fixes.

* Additional macOS audio compiler error fixes.

* Fix Windows build once and for all.

* Fix remaining uninitialized warnings.

* Fix data race in Socket.IO handling.

* We want to immediately fail when running sanitizers, not re-attempt tests.

* Revert accidental removals in macOS CI config.

* Add PR #1076 to changelog.

+++ src/rig_control/omnirig/OmniRigController.cpp
@@ -255,7 +255,7 @@ void OmniRigController::setFrequencyImpl_(uint64_t frequencyHz)
             }
         }
 
-        if (result != S_OK || tmpFreq != frequencyHz)
+        if (result != S_OK || (uint64_t)tmpFreq != frequencyHz)
         {
             // Try VFO-less set in case the first one didn't work.
             log_info("Trying frequency set fallback");
